// src/app/api/model/direccion.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Direccion { 
    id?: string;
    poblacion?: string;
    codigoPostal?: string;
    calle?: string;
    portal?: number;
    piso?: string;
}